pyproject.toml
[tool.poetry]
name = "filemanager-api"
version = "0.0.1"
description = "async, api, file_manager"
authors = ["DimaCheban <dimacheban23@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.115.8"}
pydantic = "^2.10.6"
sqlalchemy = "^2.0.37"
alembic = "^1.14.1"
pydantic-settings = "^2.7.1"
python-dotenv = "^1.0.1"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"
boto3 = {extras = ["crt"], version = "^1.36.8"}
minio = "^7.2.15"
aiofiles = "^24.1.0"
bcrypt = "^4.2.1"
pyjwt = "^2.10.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pytz = "^2024.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
types-python-jose = "^3.3.4.20240106"
types-pytz = "^2024.2.0.20241221"
aio-pika = "^9.5.4"
fastapi-mail = "^1.4.2"
pytest = "^8.3.4"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.25.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
